+++ tests/fastapi/test_fastapi.py
@@ -9,19 +9,19 @@ from fastapi import FastAPI
 from fastapi.testclient import TestClient
 
 from dockerflow import checks
-from dockerflow.fastapi import RequestIdLogFilter
 from dockerflow.fastapi import router as dockerflow_router
 from dockerflow.fastapi.middleware import (
-    CorrelationIdMiddleware,
     MozlogRequestSummaryLogger,
+    RequestIdMiddleware,
 )
+from dockerflow.logging import JsonLogFormatter, RequestIdLogFilter
 
 
 def create_app():
     app = FastAPI()
     app.include_router(dockerflow_router)
     app.add_middleware(MozlogRequestSummaryLogger)
-    app.add_middleware(CorrelationIdMiddleware, validator=None)
+    app.add_middleware(RequestIdMiddleware)
     return app
 
 
@@ -94,10 +94,8 @@ def test_mozlog_without_correlation_id_middleware(client, caplog):
 
 
 def test_request_id_passed_to_all_log_messages(caplog):
-    handler = logging.StreamHandler()
-    handler.addFilter(RequestIdLogFilter())
-    _logger = logging.getLogger("some_logger")
-    _logger.addHandler(handler)
+    caplog.handler.addFilter(RequestIdLogFilter())
+    caplog.handler.setFormatter(JsonLogFormatter())
 
     app = create_app()
 
@@ -113,6 +111,8 @@ def test_request_id_passed_to_all_log_messages(caplog):
 
     log_message = next(r for r in caplog.records if r.name == "some_logger")
     assert log_message.rid is not None
+    parsed_log = json.loads(caplog.text.splitlines()[0])
+    assert "rid" in parsed_log["Fields"]
 
 
 def test_mozlog_failure(client, mocker, caplog):